backend/style-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0 